Move the JavaScript files which will be served to client folder.
- The gulp folder is for gulp task helper only.
