+++ Nancy.ViewEngines.React/tools/client/console.js
