+++ Nancy.ViewEngines.React/tools/client/layout.jsx
