+++ Nancy.ViewEngines.React/tools/gulp/entry.js
@@ -47,7 +47,7 @@ function buildEntryCode(items) {
     ? ''
     : `if (typeof window !== "undefined") {\n${clientCode};\n}\n\n`;
 
-  const requireRender = requireLibrary('./render.js');
+  const requireRender = requireLibrary('../client/render.js');
   const lookupCode = lookup.map(formatLine).join(',\n');
   const requireLayout = requireView(options.layout);
   const renderCode = `module.exports = ${requireRender}({\n${lookupCode}\n}, ${requireLayout});`;
